move calib code from legacypipe to legacyzpts
